+++ includes/rest/class-server.php
@@ -13,6 +13,8 @@ use WP_REST_Response;
 use Activitypub\Signature;
 use Activitypub\Model\Application;
 
+use function Activitypub\use_authorized_fetch;
+
 /**
  * ActivityPub Server REST-Class.
  *
@@ -128,7 +130,7 @@ class Server {
 			// POST-Requests are always signed.
 			'GET' !== $request->get_method() ||
 			// GET-Requests only require a signature in secure mode.
-			( 'GET' === $request->get_method() && ACTIVITYPUB_AUTHORIZED_FETCH )
+			( 'GET' === $request->get_method() && use_authorized_fetch() )
 		) {
 			$verified_request = Signature::verify_http_signature( $request );
 			if ( \is_wp_error( $verified_request ) ) {